Fix usage of character ranges for methods that require glyph indexes, crashing at end of document
You can’t just pass in character ranges where glyphs are required, as
they’re not always equivalent.

Additionally, this fixes getting the max of the range and feeding it to
the line function when at the end of the doc…that char index is
nonexistent because it represents an insertion point at the end of the
text storage.